requirements.txt
streamlit>=1.36
moviepy==1.0.3
pydub>=0.25.1
python-dotenv>=1.0.1
requests>=2.32.3
openai>=1.50.0
elevenlabs>=1.7.0
srt>=3.5.3